requirements.txt
anyio==3.6.2
click==8.1.3
colorama==0.4.6
fastapi==0.85.0
h11==0.14.0
idna==3.4
kafka-python==2.0.2
pydantic==1.10.4
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.4.0
uvicorn==0.18.3
certifi==2022.12.7 
charset-normalizer==2.1.1 
requests==2.28.1 
urllib3==1.26.13
asyncio==3.4.3